Teleport range now dependent on congig file

+++ Assets/Board Game Assets/Scripts/Board Game/GameConfig/TeleportConfig.cs
@@ -1,7 +1,7 @@
 using Newtonsoft.Json;
 
 public class TeleportConfig : ConfigData {
-    public float teleportRange = 3f;
+    public int teleportRange = 3;
 
     public static TeleportConfig FromJson(string json) {
         return JsonConvert.DeserializeObject<TeleportConfig>(json);
